plAttack & ftDevice work (#1990)
* ftdevice work

* remove ftCo_800C07F8 arg2 type to be generic

* rename Player_GetTotalAttackCountPtr to Player_GetActionStats
add plStats_Attack enum

* plActionStats struct work

* plAttack_8003759C work

* More easy_funcs (#1987)

* More easy_funcs

* Adding needed prototype

* Added needed prototype

---------

+++ src/melee/gr/grgreatbay.c
@@ -205,7 +205,7 @@ void grGreatBay_801F451C(Ground_GObj* gobj)
 
 void fn_801F4520(HSD_GObj* gobj)
 {
-    ftCo_800C07F8(gobj, 6, (int) fn_801F6708);
+    ftCo_800C07F8(gobj, 6, fn_801F6708);
 }
 
 void grGreatBay_801F454C(Ground_GObj* gobj)
